+++ Tp1 Systeme de voiture/Vehicle.h
@@ -5,25 +5,27 @@
 class Vehicle
 {
 protected:
+
 	static unsigned int s_nextAvailableVehicleId;
 	unsigned int m_vehicleId;
+
 	E_Color m_color = E_Color::Count;
 	E_VehicleType m_vehicleType = E_VehicleType::Count;
 	int m_price = 0;
-	int m_maxPrice = 400000;
+	long m_maxPrice = 400000;
 	bool m_isSold = false;
 
 public:
 	Vehicle();
-	Vehicle(E_Color color, int price, bool isSold);
+	Vehicle(E_Color color, long price, bool isSold);
 
 	E_Color GetColor();
 	void SetColor(E_Color color);
 
 	const E_VehicleType GetVehicleType();
 
-	void SetPrice(int newPrice);
-	const int GetPrice();
+	void SetPrice(long newPrice);
+	const long GetPrice();
 
 	void SetIsSold(bool newIsSold);
 	const bool GetIsSold();
